Add messages to messagetracker before getting produce response
This avoids a race condition where the message would sometimes be consumed before we receive the produce response. If that's the case we would consider this message as lost, because we delete the message from the cache before it has been entered. Then after the preconfigured expiration time we would still have this message in the tracker, even though we have already read it.