fix: ensure each broker is preferred leader for a partition
The end-to-end monitoring service already includes a process to validate
that each broker is the preferred leader for at least one partition, but
this validation was not run for brokers that were not the preferred
leader for any partitions.

For example, this would mean an additional partition would never be
created after the initial creation of the management topic in the case
of another broker being added to the cluster.

This fix could be a breaking change for users without the required
permissions to create partitions or alter partition assignments if they
have inadvertently relied on this feature not working as documented.